+++ Sources/BaseFoundationiOS/Protocols/CoreProtocols/Copying.swift
@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  Copying.swift
 //  
 //
 //  Created by pablo borquez on 28/07/23.